Tools: Extracted github queries from `update-pr-testresults.js`
