+++ tools/gulptasks/update-pr-testresults.js
@@ -3,13 +3,13 @@
  */
 
 const gulp = require('gulp');
-const request = require('request');
 const fs = require('fs');
 const logLib = require('./lib/log');
 const argv = require('yargs').argv;
 const highchartsVersion = require('../../package').version;
 const { getFilesChanged, getLatestCommitShaSync } = require('./lib/git');
 const { uploadFiles, getS3Object, putS3Object } = require('./lib/uploadS3');
+const { doRequest, createPRComment, updatePRComment, fetchPRComments } = require('./lib/github');
 
 const S3_PULLREQUEST_PATH = 'visualtests/diffs/pullrequests';
 const S3_REVIEWS_PATH = 'visualtests/reviews';
@@ -25,57 +25,6 @@ const DEFAULT_OPTIONS = {
 };
 
 const VISUAL_TESTS_BUCKET = process.env.HIGHCHARTS_VISUAL_TESTS_BUCKET || 'staging-vis-dev.highcharts.com';
-
-/**
- * Executes a request with the specified options
- *
- * @param {any} options to add (see node request module)
- * @return {Promise<*> | Promise | Promise} Promise to keep
- */
-async function doRequest(options = {}) {
-    logLib.message(options.method + ' request to ' + options.url);
-    return new Promise((resolve, reject) => {
-        request(options, (error, response, data) => {
-            if (error || response.statusCode >= 400) {
-                reject(error ? error : `HTTP ${response.statusCode} - ${data.message}`);
-            } else {
-                resolve(data);
-            }
-        });
-    });
-}
-
-/**
- * Fetches the pull request comments and filters them by user and filter text.
- *
- * @param {number} pr to fetch comments for
- * @param {string} user to filter on
- * @param {string} filterText to filter on
- * @return {Promise<*> | Promise | Promise} Promise to keep
- */
-async function fetchPRComments(pr, user, filterText) {
-    if (argv.dryrun) {
-        return Promise.resolve('Dryrun (skipping fetch of PR comments)..');
-    }
-    return new Promise((resolve, reject) => {
-        doRequest({
-            ...DEFAULT_OPTIONS,
-            url: `https://api.github.com/repos/highcharts/highcharts/issues/${pr}/comments`
-        }).then(response => {
-            let comments = [];
-            if (response.length > 0) {
-                comments =
-                    response.filter(
-                        comment => comment.user.login === user && comment.body && comment.body.includes(filterText)
-                    );
-            }
-            resolve(comments);
-        })
-            .catch(err => {
-                reject(new Error(`Failed to fetch comments for PR #${pr}.: ` + err));
-            });
-    });
-}
 
 /**
  * Updates the status of the commit on github with the provided status in order to display it in pr.
@@ -130,75 +79,6 @@ async function postGitCommitStatusUpdate(pr, newReview) {
         logLib.warn(`Failed to create github status for sha ${commitSha}: ${error.message}`);
     }
     return response;
-}
-
-/**
- * Updates an existing Github comment
- *
- * @param {number} commentId to update
- * @param {string} newComment to overwrite existing one
- * @return {Promise<*> | Promise | Promise} Promise to keep
- */
-async function updatePRComment(commentId, newComment) {
-    if (argv.dryrun) {
-        logLib.message('Dryrun (skipping update of PR comment): ', newComment);
-        // eslint-disable-next-line camelcase
-        return Promise.resolve({ html_url: 'No where' });
-    }
-
-    logLib.message('Updating existing comment with id ' + commentId);
-    return new Promise((resolve, reject) => {
-        doRequest({
-            ...DEFAULT_OPTIONS,
-            url: `https://api.github.com/repos/highcharts/highcharts/issues/comments/${commentId}`,
-            method: 'PATCH',
-            body: { body: newComment }
-        })
-            .then(response => {
-                logLib.message(`Comment updated at ${response.html_url}`);
-
-                resolve(response);
-            })
-            .catch(err => {
-                logLib.warn('Failed to update existing PR comment: ' + err);
-                reject(err);
-            });
-    });
-}
-
-/**
- * Creates a Github PR comment
- * @param {number} pr number
- * @param {string} comment to post
- * @return {Promise<*> | Promise | Promise} Promise to keep
- */
-async function createPRComment(pr, comment) {
-    if (argv.dryrun) {
-        logLib.message('(Dryrun) Skipping creation of pr comment: ', comment);
-        // eslint-disable-next-line camelcase
-        return Promise.resolve({ html_url: 'No where' });
-    }
-
-    return new Promise((resolve, reject) => {
-        doRequest({
-            ...DEFAULT_OPTIONS,
-            url: `https://api.github.com/repos/highcharts/highcharts/issues/${pr}/comments`,
-            method: 'POST',
-            body: { body: comment }
-        })
-            .then(result => {
-                logLib.message(`Comment created at ${result.html_url}`);
-                resolve(result);
-            })
-            .catch(err => {
-                const failureMsg = 'Failed to create PR comment: ' + err;
-                logLib.warn(failureMsg);
-                if (argv.failSilently) {
-                    resolve(failureMsg);
-                }
-                reject(new Error(failureMsg));
-            });
-    });
 }
 
 /**
